refactor(solver): thread subtype eval session state (#14346)
