+++ crates/tsz-checker/src/query_boundaries/assignability/overload_subtype_pass.rs
@@ -44,6 +44,7 @@ pub(crate) fn cached_overload_subtype_pass_assignability<
 
     let context = tsz_solver::relations::relation_queries::RelationContext {
         query_db: Some(inputs.db),
+        evaluation_session: None,
         inheritance_graph: Some(inputs.inheritance_graph),
         class_check: None,
     };
