+++ crates/tsz-solver/src/relations/subtype/rules/functions/checking.rs
@@ -12,10 +12,12 @@ use crate::types::{
 };
 use crate::visitor::callable_shape_id;
 
-use super::super::super::{RelationEvaluationResult, SubtypeChecker, SubtypeResult, TypeResolver};
+use super::super::super::{SubtypeChecker, SubtypeResult, TypeResolver};
 use super::erase_type_params_to_constraints;
 
+mod call_signatures;
 mod context_instantiation;
+mod evaluation;
 mod generic_constraints;
 mod name_pairing;
 mod overloads;
@@ -1960,239 +1962,5 @@ impl<'a, R: TypeResolver> SubtypeChecker<'a, R> {
                     self.effective_param_type_pair(source_param, target_param);
                 self.are_parameters_compatible_impl(source_type, target_type, true)
             })
-    }
-
-    /// Check call signature subtyping.
-    pub(crate) fn check_call_signature_subtype(
-        &mut self,
-        source: &CallSignature,
-        target: &CallSignature,
-    ) -> SubtypeResult {
-        self.check_call_signature_subtype_impl(source, target, false)
-    }
-
-    pub(crate) fn callable_modality_flags_for_type(&mut self, type_id: TypeId) -> (bool, bool) {
-        let direct = self.callable_modality_flags_for_type_direct(type_id);
-        if direct.0 || direct.1 {
-            return direct;
-        }
-        let evaluated = self.evaluate_type(type_id);
-        if evaluated == type_id {
-            direct
-        } else {
-            self.callable_modality_flags_for_type_direct(evaluated)
-        }
-    }
-
-    fn callable_modality_flags_for_type_direct(&self, type_id: TypeId) -> (bool, bool) {
-        if let Some(shape_id) = callable_shape_id(self.interner, type_id) {
-            let shape = self.interner.callable_shape(shape_id);
-            return (
-                !shape.call_signatures.is_empty(),
-                !shape.construct_signatures.is_empty(),
-            );
-        }
-        if let Some(fn_id) = crate::visitor::function_shape_id(self.interner, type_id) {
-            let f = self.interner.function_shape(fn_id);
-            return (!f.is_constructor, f.is_constructor);
-        }
-        (false, false)
-    }
-
-    pub(crate) fn check_call_signature_subtype_as_constructor(
-        &mut self,
-        source: &CallSignature,
-        target: &CallSignature,
-    ) -> SubtypeResult {
-        self.check_call_signature_subtype_impl(source, target, true)
-    }
-
-    fn check_call_signature_subtype_impl(
-        &mut self,
-        source: &CallSignature,
-        target: &CallSignature,
-        is_constructor: bool,
-    ) -> SubtypeResult {
-        let source_fn = FunctionShape {
-            type_params: source.type_params.clone(),
-            params: source.params.clone(),
-            this_type: source.this_type,
-            return_type: source.return_type,
-            type_predicate: source.type_predicate,
-            is_constructor,
-            is_method: source.is_method,
-        };
-        let target_fn = FunctionShape {
-            type_params: target.type_params.clone(),
-            params: target.params.clone(),
-            this_type: target.this_type,
-            return_type: target.return_type,
-            type_predicate: target.type_predicate,
-            is_constructor,
-            is_method: target.is_method,
-        };
-        self.check_function_subtype(&source_fn, &target_fn)
-    }
-
-    fn constructor_signatures_need_strict_params(
-        source: &FunctionShape,
-        target: &FunctionShape,
-    ) -> bool {
-        if !(source.is_constructor || target.is_constructor) {
-            return false;
-        }
-
-        let source_generic = !source.type_params.is_empty();
-        let target_generic = !target.type_params.is_empty();
-        if !source_generic && !target_generic {
-            // Non-generic constructors need strict params when there's an
-            // optionality mismatch between corresponding parameters. This
-            // matches tsc where property-typed constructor types like
-            // `new (x?: number) => number` use strict comparison, not
-            // constructor bivariance. Without this, the bivariant check
-            // would pass (`number <: number | undefined`) and incorrectly
-            // allow `new (x: number) => number` as a subtype.
-            let has_optionality_mismatch = source
-                .params
-                .iter()
-                .zip(target.params.iter())
-                .any(|(sp, tp)| sp.optional != tp.optional);
-            return has_optionality_mismatch;
-        }
-
-        if source_generic && !target_generic {
-            let has_optionality_mismatch = source
-                .params
-                .iter()
-                .zip(target.params.iter())
-                .any(|(sp, tp)| sp.optional != tp.optional);
-            return has_optionality_mismatch
-                || source.type_params.iter().any(|tp| tp.constraint.is_some());
-        }
-
-        source.type_params.len() != target.type_params.len()
-            || source
-                .type_params
-                .iter()
-                .chain(target.type_params.iter())
-                .any(|tp| tp.constraint.is_some())
-            || source.params.len() != 1
-            || target.params.len() != 1
-    }
-
-    /// Check call signature subtype to function shape.
-    pub(crate) fn check_call_signature_subtype_to_fn(
-        &mut self,
-        source: &CallSignature,
-        target: &FunctionShape,
-    ) -> SubtypeResult {
-        let source_fn = FunctionShape {
-            type_params: source.type_params.clone(),
-            params: source.params.clone(),
-            this_type: source.this_type,
-            return_type: source.return_type,
-            type_predicate: source.type_predicate,
-            is_constructor: target.is_constructor,
-            is_method: source.is_method,
-        };
-        self.check_function_subtype(&source_fn, target)
-    }
-
-    /// Check function shape subtype to call signature.
-    pub(crate) fn check_call_signature_subtype_fn(
-        &mut self,
-        source: &FunctionShape,
-        target: &CallSignature,
-    ) -> SubtypeResult {
-        let target_fn = FunctionShape {
-            type_params: target.type_params.clone(),
-            params: target.params.clone(),
-            this_type: target.this_type,
-            return_type: target.return_type,
-            type_predicate: target.type_predicate,
-            is_constructor: source.is_constructor,
-            is_method: target.is_method,
-        };
-        self.check_function_subtype(source, &target_fn)
-    }
-
-    /// Evaluate a meta-type (conditional, index access, mapped, keyof, etc.) to its
-    /// concrete form. Uses `TypeEvaluator` with the resolver to correctly resolve
-    /// Lazy(DefId) types at all nesting levels (e.g., KeyOf(Lazy(DefId))).
-    ///
-    /// Always uses `TypeEvaluator` with the resolver instead of `query_db.evaluate_type()`
-    /// because the checker populates DefId→TypeId mappings in the `TypeEnvironment` that
-    /// the `query_db`'s resolver-less evaluator cannot access.
-    ///
-    /// Results are cached in `eval_cache` to avoid re-evaluating the same type across
-    /// multiple subtype checks. This turns O(n²) evaluate calls into O(n).
-    pub(crate) fn evaluate_type(&mut self, type_id: TypeId) -> TypeId {
-        self.evaluate_type_with_stability(type_id).type_id()
-    }
-
-    /// Like [`Self::evaluate_type`], but also reports whether the evaluation is
-    /// *stable*: it converged without tripping any recursion/depth/budget limit
-    /// or cross-instance cycle bail.
-    ///
-    /// Callers that special-case collapsed `unknown` results must consult the
-    /// flag so a genuine `unknown` is not treated like a bail artifact.
-    pub(crate) fn evaluate_type_with_stability(
-        &mut self,
-        type_id: TypeId,
-    ) -> RelationEvaluationResult {
-        // Fast path: intrinsic types (number, string, boolean, void, null, etc.)
-        // never need evaluation. Skip cache lookup entirely.
-        if type_id.is_intrinsic() {
-            return RelationEvaluationResult::stable(type_id);
-        }
-        // Check local evaluation cache first.
-        // Key includes no_unchecked_indexed_access since with that flag evaluation results can vary.
-        let cache_key = (type_id, self.no_unchecked_indexed_access);
-        if let Some(&cached) = self.eval_cache.get(&cache_key) {
-            return cached;
-        }
-        // Charge the failure-explanation work budget (#13243) for each distinct
-        // (cache-miss) evaluation. Cache hits above are free, so this counts the
-        // sub-types the explain pass actually drills into. Outside the explain
-        // pass the fuel is `None` and this is inert. When the budget is
-        // exhausted, return the type unevaluated and flagged unstable: the
-        // explain traversal then collapses to a coarse `TypeMismatch` (see
-        // `explain_failure_guarded`) instead of driving another combinatorial
-        // `instantiate_and_finalize_application`.
-        if let Some(fuel) = self.explain_eval_fuel.as_mut() {
-            if *fuel == 0 {
-                return RelationEvaluationResult::unstable(type_id);
-            }
-            *fuel -= 1;
-        }
-        use crate::evaluation::cross_eval_guard;
-        use crate::evaluation::evaluate::TypeEvaluator;
-        // Per-query memo + cross-instance cycle break (#11586): the subtype checker
-        // spins up a fresh `TypeEvaluator` here whose per-instance recursion guard
-        // starts empty, so a recursive conditional/`infer` application related
-        // against a structural target can bounce evaluator → subtype-checker →
-        // evaluator, re-entering the same `TypeId` without any guard firing. On a
-        // cross-instance cycle (`None`) leave the type unevaluated and, crucially,
-        // do not record it in `eval_cache` — the in-flight ancestor owns the result.
-        let Some(memo_result) = cross_eval_guard::memoized_eval_with_stability(
-            type_id,
-            self.no_unchecked_indexed_access,
-            || {
-                let mut evaluator = TypeEvaluator::with_resolver(self.interner, self.resolver);
-                // Pass query_db to share the application evaluation cache across
-                // evaluations, keeping structural subtype checks stable.
-                if let Some(db) = self.query_db {
-                    evaluator = evaluator.with_query_db(db);
-                }
-                let request = crate::evaluation::request::EvaluationRequest::new(type_id)
-                    .with_no_unchecked_indexed_access(self.no_unchecked_indexed_access);
-                evaluator.evaluate_request_memo_result(request)
-            },
-        ) else {
-            return RelationEvaluationResult::unstable(type_id);
-        };
-        let entry = RelationEvaluationResult::from_depth_agnostic_memo(memo_result);
-        self.eval_cache.insert(cache_key, entry);
-        entry
     }
 }
